backend/pyproject.toml
[tool.poetry]
name = "backend service"
version = "0.1.0"
description = ""
authors = ["Jonathan Alberth Quispe Fuentes <qf.jonathan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-extensions = "^3.1.5"
Werkzeug = "^2.0.2"
graphene-django = "^2.15.0"
django-graphql-jwt = "^0.3.4"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=100
skip-string-normalization=true
target-version=['py39']

[flake8]
max-line-length=100
